requirements.txt
google-api-python-client==2.146.0
google-auth==2.33.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
PyYAML==6.0.2
schedule==1.2.2
streamlit==1.38.0
